Ajout Token CSRF

+++ moviedb/src/Controller/Admin/DepartmentController.php
@@ -86,10 +86,20 @@ class DepartmentController extends AbstractController
         $id = $request->request->get('department_id');
         $department = $this->getDoctrine()->getRepository(Department::class)->find($id);
 
-        $em = $this->getDoctrine()->getManager();
-        // la méthode remove de l'entityManager permet de supprimer l'objet de la BDD
-        $em->remove($department);
-        $em->flush();
+        //Récupération du jeton dans la requête
+        $token = $request->request->get('_token');
+        // Vérification du jeton CSRF (il permet de vérifier que les infos reçues viennent bien d'un formulaire de Symfony)
+        if ($this->isCsrfTokenValid('delete-department', $token)){
+            $em = $this->getDoctrine()->getManager();
+            // la méthode remove de l'entityManager permet de supprimer l'objet de la BDD
+            $em->remove($department);
+            $em->flush();
+
+            $this->addFlash('success', 'Le Département a bien été supprimé !');
+        } else {
+            $this->addFlash('danger', 'Votre formulaire est invalide, veuillez recommencer !');
+        }
+
 
         //Une fois fini, on redirige vers la liste des departments
         return $this->redirectToRoute('admin_department');
